pyproject.toml
[project]
name = "venmo-auto-cashout"
version = "1.0"
description = "A tool to post Venmo transactions to Lunch Money"
authors = [
    {name = "Kahlil Oppenheimer", email = "unrealkahlil@gmail.com"},
]
dependencies = [
    "lunchable~=0.1.11",
    "venmo-api-unofficial @ https://github.com/evanpurkhiser/venmo-api-unofficial/archive/13bf98aa9b12f4b48f3326028bfd12dc5dfbbdce.tar.gz",
    "sentry-sdk~=1.5",
    "requests~=2.27",
    "pyotp~=2.6",
]
requires-python = ">=3.9"
dynamic = ["classifiers"]
license = {text = "MIT"}

[project.urls]
homepage = "https://github.com/evanpurkhiser/venmo-auto-cashout"

[project.scripts]
venmo-auto-cashout = "venmo_auto_cashout:main"

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"

[tool.black]
line-length=100
target-version=['py39']

[tool.isort]
profile = "black"
line_length = 100
lines_between_sections = 1
skip = "migrations"

[tool.pdm]

[tool.pyright]
include = ["venmo_auto_cashout"]
extraPaths = ["__pypackages__/3.9/lib"]
